chore(cd): add integration test
