+++ cd/main_test.go
@@ -1,371 +1,158 @@
+//go:build integration
+
 package main
 
 import (
+	"bytes"
 	"compress/gzip"
 	"encoding/json"
-	"net/http"
-	"net/http/httptest"
+	"fmt"
+	"io"
+	"os"
+	"os/exec"
+	"regexp"
+	"slices"
+	"strings"
 	"testing"
 
 	"github.com/pulumi/pulumi/sdk/v3/go/auto/events"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
+	"github.com/stretchr/testify/require"
 )
 
-func TestColor(t *testing.T) {
-	saved := noColor
-	t.Cleanup(func() { noColor = saved })
-
-	noColor = false
-	if got := color(); got != "always" {
-		t.Errorf("color() = %q, want %q", got, "always")
-	}
-
-	noColor = true
-	if got := color(); got != "never" {
-		t.Errorf("color() = %q, want %q", got, "never")
+func TestRunVersion(t *testing.T) {
+	if run("cd", "--version") != 0 {
+		t.Error("run --version flag should exit with 0")
 	}
 }
 
-func TestProjectConfig(t *testing.T) {
-	cfg := projectConfig("TestPrefix")
-
-	autonaming, ok := cfg["pulumi:autonaming"]
-	if !ok {
-		t.Fatal("missing pulumi:autonaming config")
-	}
-	autonamingMap, ok := autonaming.Value.(map[string]any)
-	if !ok {
-		t.Fatal("autonaming value is not map[string]any")
-	}
-	pattern, ok := autonamingMap["pattern"].(string)
-	if !ok {
-		t.Fatal("pattern is not a string")
-	}
-	if pattern != "TestPrefix-${project}-${stack}-${name}-${hex(7)}" {
-		t.Errorf("unexpected pattern: %s", pattern)
-	}
-
-	disabled, ok := cfg["pulumi:disable-default-providers"]
-	if !ok {
-		t.Fatal("missing pulumi:disable-default-providers config")
-	}
-	disabledList, ok := disabled.Value.([]string)
-	if !ok {
-		t.Fatal("disable-default-providers value is not []string")
-	}
-	if len(disabledList) < 3 {
-		t.Errorf("expected >= 3 disabled provider, got %d", len(disabledList))
+func TestRunUsageError(t *testing.T) {
+	if run() != 2 {
+		t.Error("run should exit with 2 on usage error")
 	}
 }
 
-func TestProjectConfigEmptyPrefix(t *testing.T) {
-	cfg := projectConfig("")
+func TestRunPreviewAzure(t *testing.T) {
+	t.Setenv("AZURE_SUBSCRIPTION_ID", "f311c4db-e998-4c94-906c-7e2637303a05")
+	t.Setenv("AZURE_LOCATION", "westus")
 
-	autonaming := cfg["pulumi:autonaming"]
-	autonamingMap := autonaming.Value.(map[string]any)
-	pattern := autonamingMap["pattern"].(string)
-	// With empty prefix, pattern should not have a leading prefix-
-	if pattern != "${project}-${stack}-${name}-${hex(7)}" {
-		t.Errorf("unexpected pattern with empty prefix: %s", pattern)
-	}
+	testProviderPreview(t, "azure")
 }
 
-func TestGetenv(t *testing.T) {
-	tests := []struct {
-		name     string
-		key      string
-		envVal   string
-		fallback string
-		want     string
-	}{
-		{"env set", "TEST_GETENV_SET", "fromenv", "fallback", "fromenv"},
-		{"env empty", "TEST_GETENV_EMPTY", "", "fallback", "fallback"},
-		{"env unset", "TEST_GETENV_UNSET_XYZ", "", "fallback", "fallback"},
-	}
+func TestRunPreviewAWS(t *testing.T) {
+	t.Setenv("AWS_PROFILE", "defang-lab")
+	t.Setenv("AWS_REGION", "us-west-2")
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.envVal != "" {
-				t.Setenv(tt.key, tt.envVal)
-			}
-			if got := getenv(tt.key, tt.fallback); got != tt.want {
-				t.Errorf("Getenv(%q, %q) = %q, want %q", tt.key, tt.fallback, got, tt.want)
-			}
-		})
-	}
+	testProviderPreview(t, "aws")
 }
 
-func TestSplitByComma(t *testing.T) {
-	tests := []struct {
-		input string
-		want  []string
-	}{
-		{"", nil},
-		{"a", []string{"a"}},
-		{"a,b,c", []string{"a", "b", "c"}},
-		{"one,,three", []string{"one", "", "three"}},
-	}
+func TestRunPreviewGCP(t *testing.T) {
+	t.Setenv("GCP_PROJECT", "liotest-443018") // TODO: pick a neutral project for integration testing
+	t.Setenv("GCP_REGION", "us-central1")
 
-	for _, tt := range tests {
-		t.Run(tt.input, func(t *testing.T) {
-			got := splitByComma(tt.input)
-			if tt.want == nil {
-				if got != nil {
-					t.Errorf("SplitByComma(%q) = %v, want nil", tt.input, got)
-				}
-				return
-			}
-			if len(got) != len(tt.want) {
-				t.Fatalf("SplitByComma(%q) = %v, want %v", tt.input, got, tt.want)
-			}
-			for i := range got {
-				if got[i] != tt.want[i] {
-					t.Errorf("SplitByComma(%q)[%d] = %q, want %q", tt.input, i, got[i], tt.want[i])
-				}
-			}
-		})
-	}
+	testProviderPreview(t, "gcp")
 }
 
-func TestGetenvBool(t *testing.T) {
-	tests := []struct {
-		name string
-		val  string
-		want bool
-	}{
-		{"true", "true", true},
-		{"1", "1", true},
-		{"false", "false", false},
-		{"0", "0", false},
-		{"empty", "", false},
+func testProviderPreview(t *testing.T, provider string) {
+	t.Helper()
+
+	if err := exec.CommandContext(t.Context(), "make", "-C", "..", "install_defang-"+provider).Run(); err != nil {
+		t.Fatalf("failed to install defang-%s provider: %v", provider, err)
 	}
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			key := "TEST_GETENVBOOL_" + tt.name
-			t.Setenv(key, tt.val)
-			if got := getenvBool(key); got != tt.want {
-				t.Errorf("GetenvBool(%q) with value %q = %v, want %v", key, tt.val, got, tt.want)
-			}
-		})
-	}
-}
+	eventsFile := t.TempDir() + "/events.json.gz"
+	t.Setenv("DEFANG_EVENTS_UPLOAD_URL", "file://"+eventsFile)
+	statesFile := t.TempDir() + "/states.json.gz"
+	t.Setenv("DEFANG_STATES_UPLOAD_URL", "file://"+statesFile)
+	t.Setenv("DEFANG_PULUMI_DIFF", "true")
+	t.Setenv("DEFANG_PULUMI_DEBUG", "false")
+	t.Setenv("NO_COLOR", "")
+	t.Setenv("PROJECT", "cd-test")
+	t.Setenv("STACK", provider)
 
-func TestStackConfigAWS(t *testing.T) {
-	// Save and restore all globals that stackConfig reads
-	savedAwsRegion := awsRegion
-	savedGcpProjectId := gcpProjectId
-	savedAzureSubscriptionId := azureSubscriptionId
-	savedOrg := org
-	savedPrefix := prefix
-	savedMode := mode
-	savedDomain := domain
-	savedPrivateDomain := privateDomain
-	savedDelegationSetId := delegationSetId
-	savedRegistryCredsArn := registryCredsArn
-	savedAwsProfile := awsProfile
-	savedStateUrl := stateUrl
-	t.Cleanup(func() {
-		awsRegion = savedAwsRegion
-		gcpProjectId = savedGcpProjectId
-		azureSubscriptionId = savedAzureSubscriptionId
-		org = savedOrg
-		prefix = savedPrefix
-		mode = savedMode
-		domain = savedDomain
-		privateDomain = savedPrivateDomain
-		delegationSetId = savedDelegationSetId
-		registryCredsArn = savedRegistryCredsArn
-		awsProfile = savedAwsProfile
-		stateUrl = savedStateUrl
+	if run("cd", "preview", "IlpzZXJ2aWNlczoKICBuZ2lueDoKICAgIGltYWdlOiBuZ2lueAogICAgcG9ydHM6CiAgICAgIC0gdGFyZ2V0OiA4MAogICAgICAgIHB1Ymxpc2hlZDogIjgwIgo=") != 0 {
+		t.Fatal("preview should exit with 0")
+	}
+
+	// Ensure events were uploaded
+	uploaded, err := readFile(eventsFile)
+	if err != nil {
+		t.Fatalf("failed to load events: %v", err)
+	}
+
+	// Normalize generated random prefixes. The 4-7 hex suffix from autonaming
+	// (`${name}-${hex(N)}`) appears in two contexts: at the end of a JSON
+	// string (followed by `"`) and embedded in GCP service-account emails
+	// (followed by `@`). Match both so the suffix gets erased in either spot.
+	uploaded = regexp.MustCompile(`-[0-9a-f]{4,7}(["@])`).ReplaceAll(uploaded, []byte(`-***$1`))
+
+	var eventJson struct {
+		Events []events.EngineEvent `json:"events"`
+	}
+	if err := json.Unmarshal(uploaded, &eventJson); err != nil {
+		t.Fatalf("failed to unmarshal events: %v", err)
+	}
+
+	// Drop debug-severity diagnostics: these are Pulumi-engine RPC firehose
+	// logs ("RegisterResource RPC prepared/finished: ...") emitted in
+	// concurrent order with no URN to sort by — pure non-determinism noise
+	// for a golden snapshot.
+	eventJson.Events = slices.DeleteFunc(eventJson.Events, func(e events.EngineEvent) bool {
+		return e.DiagnosticEvent != nil && e.DiagnosticEvent.Severity == "debug"
 	})
 
-	awsRegion = "us-east-1"
-	gcpProjectId = ""
-	azureSubscriptionId = ""
-	org = "testorg"
-	prefix = "TestPrefix"
-	mode = "development"
-	domain = "example.com"
-	privateDomain = "internal.example.com"
-	delegationSetId = "DELEGSET123"
-	registryCredsArn = "arn:aws:secretsmanager:us-east-1:123456789:secret:creds"
-	awsProfile = "myprofile"
-	stateUrl = "http://example.com/state"
-
-	cfg, err := stackConfig()
-	if err != nil {
-		t.Fatalf("stackConfig() error: %v", err)
+	// Normalize timestamps, sequence numbers, duration
+	for i, e := range eventJson.Events {
+		eventJson.Events[i].Timestamp = 42
+		eventJson.Events[i].Sequence = 42
+		if e.SummaryEvent != nil {
+			e.SummaryEvent.DurationSeconds = 42
+		}
 	}
-
-	if cfg["defang:provider"].Value != "aws" {
-		t.Errorf("expected provider aws, got %q", cfg["defang:provider"].Value)
+	// Stable sort by URN across all resource-event variants. Non-resource
+	// events (Summary, Prelude, Diagnostic, ...) get an empty URN and stay
+	// in their original relative order thanks to SortStable.
+	urn := func(e events.EngineEvent) string {
+		switch {
+		case e.ResourcePreEvent != nil:
+			return e.ResourcePreEvent.Metadata.URN
+		case e.ResOutputsEvent != nil:
+			return e.ResOutputsEvent.Metadata.URN
+		case e.ResOpFailedEvent != nil:
+			return e.ResOpFailedEvent.Metadata.URN
+		}
+		return ""
 	}
-	if cfg["defang:stateUrl"].Value != "http://example.com/state" {
-		t.Errorf("expected stateUrl http://example.com/state, got %q", cfg["defang:stateUrl"].Value)
-	}
-	if cfg["aws:region"].Value != "us-east-1" {
-		t.Errorf("expected region us-east-1, got %q", cfg["aws:region"].Value)
-	}
-	if cfg["aws:profile"].Value != "myprofile" {
-		t.Errorf("expected profile myprofile, got %q", cfg["aws:profile"].Value)
-	}
-	if cfg["defang:org"].Value != "testorg" {
-		t.Errorf("expected org testorg, got %q", cfg["defang:org"].Value)
-	}
-	if cfg["defang:domain"].Value != "example.com" {
-		t.Errorf("expected domain example.com, got %q", cfg["defang:domain"].Value)
-	}
-	if cfg["defang:privateDomain"].Value != "internal.example.com" {
-		t.Errorf("expected privateDomain, got %q", cfg["defang:privateDomain"].Value)
-	}
-	if cfg["defang:delegationSetId"].Value != "DELEGSET123" {
-		t.Errorf("expected delegationSetId, got %q", cfg["defang:delegationSetId"].Value)
-	}
-	if cfg["defang:ciRegistryCredentialsArn"].Value != registryCredsArn {
-		t.Errorf("expected ciRegistryCredentialsArn, got %q", cfg["defang:ciRegistryCredentialsArn"].Value)
-	}
-}
-
-func TestStackConfigGCP(t *testing.T) {
-	savedAwsRegion := awsRegion
-	savedGcpProjectId := gcpProjectId
-	savedAzureSubscriptionId := azureSubscriptionId
-	savedOrg := org
-	savedPrefix := prefix
-	savedMode := mode
-	savedGcpRegion := gcpRegion
-	savedDomain := domain
-	savedPrivateDomain := privateDomain
-	savedDelegationSetId := delegationSetId
-	savedRegistryCredsArn := registryCredsArn
-	t.Cleanup(func() {
-		awsRegion = savedAwsRegion
-		gcpProjectId = savedGcpProjectId
-		azureSubscriptionId = savedAzureSubscriptionId
-		org = savedOrg
-		prefix = savedPrefix
-		mode = savedMode
-		gcpRegion = savedGcpRegion
-		domain = savedDomain
-		privateDomain = savedPrivateDomain
-		delegationSetId = savedDelegationSetId
-		registryCredsArn = savedRegistryCredsArn
+	slices.SortStableFunc(eventJson.Events, func(a, b events.EngineEvent) int {
+		return strings.Compare(urn(a), urn(b))
 	})
 
-	awsRegion = ""
-	gcpProjectId = "my-gcp-project"
-	azureSubscriptionId = ""
-	gcpRegion = "us-central1"
-	org = "testorg"
-	prefix = "Test"
-	mode = "production"
-	domain = ""
-	privateDomain = ""
-	delegationSetId = ""
-	registryCredsArn = ""
-
-	cfg, err := stackConfig()
+	raw, err := json.MarshalIndent(eventJson, "", "  ")
 	if err != nil {
-		t.Fatalf("stackConfig() error: %v", err)
+		t.Fatalf("failed to marshal events: %v", err)
 	}
 
-	if cfg["defang:provider"].Value != "gcp" {
-		t.Errorf("expected provider gcp, got %q", cfg["defang:provider"].Value)
+	goldenFile := "testdata/preview-events-" + provider + ".json"
+	golden, err := readFile(goldenFile)
+	if os.IsNotExist(err) || os.Getenv("UPDATE_GOLDEN") != "" {
+		// Write golden file if it doesn't exist or when UPDATE_GOLDEN is set
+		t.Fatalf("updated golden file: %v", os.WriteFile(goldenFile, raw, 0644))
+	} else if err != nil {
+		t.Fatalf("failed to load golden events: %v", err)
 	}
-	if cfg["gcp:project"].Value != "my-gcp-project" {
-		t.Errorf("expected gcp:project, got %q", cfg["gcp:project"].Value)
-	}
-	if cfg["gcp:region"].Value != "us-central1" {
-		t.Errorf("expected gcp:region, got %q", cfg["gcp:region"].Value)
-	}
-	// Optional fields should not be present when empty
-	if _, ok := cfg["defang:domain"]; ok {
-		t.Error("defang:domain should not be set when empty")
-	}
-	if _, ok := cfg["defang:privateDomain"]; ok {
-		t.Error("defang:privateDomain should not be set when empty")
-	}
+	require.JSONEq(t, string(golden), string(raw))
 }
 
-func TestStackConfigAzure(t *testing.T) {
-	savedAwsRegion := awsRegion
-	savedGcpProjectId := gcpProjectId
-	savedAzureSubscriptionId := azureSubscriptionId
-	savedAzureLocation := azureLocation
-	savedOrg := org
-	savedPrefix := prefix
-	savedMode := mode
-	savedDomain := domain
-	savedPrivateDomain := privateDomain
-	savedDelegationSetId := delegationSetId
-	savedRegistryCredsArn := registryCredsArn
-	t.Cleanup(func() {
-		awsRegion = savedAwsRegion
-		gcpProjectId = savedGcpProjectId
-		azureSubscriptionId = savedAzureSubscriptionId
-		azureLocation = savedAzureLocation
-		org = savedOrg
-		prefix = savedPrefix
-		mode = savedMode
-		domain = savedDomain
-		privateDomain = savedPrivateDomain
-		delegationSetId = savedDelegationSetId
-		registryCredsArn = savedRegistryCredsArn
-	})
-
-	awsRegion = ""
-	gcpProjectId = ""
-	azureSubscriptionId = "sub-123"
-	azureLocation = "westus2"
-	org = "testorg"
-	prefix = "Test"
-	mode = "development"
-	domain = ""
-	privateDomain = ""
-	delegationSetId = ""
-	registryCredsArn = ""
-
-	cfg, err := stackConfig()
-	if err != nil {
-		t.Fatalf("stackConfig() error: %v", err)
-	}
-
-	if cfg["defang:provider"].Value != "azure" {
-		t.Errorf("expected provider azure, got %q", cfg["defang:provider"].Value)
-	}
-	if cfg["azure-native:location"].Value != "westus2" {
-		t.Errorf("expected location westus2, got %q", cfg["azure-native:location"].Value)
-	}
-	if cfg["azure-native:useMsi"].Value != "true" {
-		t.Errorf("expected useMsi true, got %q", cfg["azure-native:useMsi"].Value)
-	}
-	if cfg["azure-native:subscriptionId"].Value != "sub-123" {
-		t.Errorf("expected subscriptionId, got %q", cfg["azure-native:subscriptionId"].Value)
-	}
-}
-
-func TestCollectEvents(t *testing.T) {
-	savedEventsUploadUrl := eventsUploadUrl
-	t.Cleanup(func() { eventsUploadUrl = savedEventsUploadUrl })
-
-	var engineEvents map[string][]events.EngineEvent
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		gz, _ := gzip.NewReader(r.Body)
+func readFile(path string) ([]byte, error) {
+	if b, err := os.ReadFile(path); err != nil {
+		return nil, err
+	} else if strings.HasSuffix(path, ".gz") {
+		gz, err := gzip.NewReader(bytes.NewReader(b))
+		if err != nil {
+			return nil, fmt.Errorf("readFile: %w", err)
+		}
 		defer gz.Close()
-		json.NewDecoder(gz).Decode(&engineEvents)
-		w.WriteHeader(http.StatusOK)
-	}))
-	t.Cleanup(srv.Close)
-
-	eventsUploadUrl = srv.URL
-	ch, wait := collectEvents(t.Context())
-	ch <- events.EngineEvent{EngineEvent: apitype.EngineEvent{Sequence: 1}}
-	ch <- events.EngineEvent{EngineEvent: apitype.EngineEvent{Sequence: 2}}
-	close(ch)
-	wait()
-
-	if len(engineEvents["events"]) != 2 {
-		t.Errorf("expected 2 events, got %d", len(engineEvents["events"]))
+		return io.ReadAll(gz)
+	} else {
+		return b, nil
 	}
 }
